colourful-constants/code/requirements.txt
colour
numpy
pyyaml
